+++ js/downloadbutton/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Dependency updates
 
+- @gradio/client@2.5.1
+
+## 0.5.2
+
+### Dependency updates
+
+- @gradio/client@2.5.0
+
+## 0.5.2
+
+### Dependency updates
+
 - @gradio/client@2.4.0
 - @gradio/utils@0.14.0
 - @gradio/button@0.8.2